tests/wpt/tests/css/css-color/parsing/color-invalid-contrast-color-function.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Test CSS Color Module 5 contrast-color() computed value</title>
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<link rel="help" href="https://drafts.csswg.org/css-color-5/#contrast-color">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
    color: pink;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>

test_invalid_value("background-color", "contrast-color()");
test_invalid_value("background-color", "contrast-color(1)");
test_invalid_value("background-color", "contrast-color(max)");
test_invalid_value("background-color", "contrast-color(max max)");
test_invalid_value("background-color", "contrast-color(max white)");
test_invalid_value("background-color", "contrast-color(white white)");
test_invalid_value("background-color", "contrast-color(white max)");
test_invalid_value("background-color", "contrast-color(white min)");
test_invalid_value("background-color", "contrast-color(white max bar)");

</script>
</body>
</html>
